server/pyproject.toml
[project]
name = "py-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-openai>=0.3.25",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.34.3",
]
